+++ config/webpack.config.js/client.prod.js
@@ -1,9 +1,10 @@
 const baseConfig = require('./client.base');
+const generateSourceMap = process.env.OMIT_SOURCEMAP === 'true' ? false : true;
 
 const config = {
     ...baseConfig,
     mode: 'production',
-    devtool: 'source-map',
+    devtool: generateSourceMap ? 'source-map' : false,
 };
 
 config.output.filename = 'bundle.[hash:8].js';